clients

+++ resources/views/roles/index.blade.php
@@ -4,7 +4,7 @@
         <div class="card-body">
             <div class="card-title mr-3 col-xs">
                 @can('edit role')
-                <a class="dropdown-item" href="{{ route('roles.edit',$role->id) }}" data-remote="true" data-type="text">
+                <a href="{{ route('roles.edit',$role->id) }}" data-remote="true" data-type="text">
                 @endcan
                 <h6 data-filter-by="text">{{ $role->name }}</h6>
                 @can('edit role')
@@ -29,12 +29,12 @@
                         <span>{{__('Edit')}}</span>
                     </a>
                     @endcan
-                    <div class="dropdown-divider"></div>
-                    @can('delete role')
+                    @if(Gate::check('delete role') && $role->name != 'client')
+                        <div class="dropdown-divider"></div>
                         <a class="dropdown-item text-danger" href="{{ route('roles.destroy', $role->id) }}" data-method="delete" data-remote="true" data-type="text">
                             <span>{{__('Delete')}}</span>
                         </a>
-                    @endcan
+                    @endif
                 </div>
             </div>
         </div>
